update DismissibleBanner component and story
fix name misspelling
